Move format to jank::runtime
